+++ src/components/Bus/BusModal/BusPassenger/BusPassengerItem/index.tsx
@@ -1,6 +1,6 @@
 import { Table, TBody, TD, TH, THead, TR } from "@b1nd/b1nd-dodamdodam-ui";
-import { BusPassengerType } from "../../../../../types/Bus/Bus.type";
-import { addPhoneHyphen } from "../../../../../utils/common/addPhoneHyphen";
+import { BusPassengerType } from "types/Bus/Bus.type";
+import { addPhoneHyphen } from "utils/common/addPhoneHyphen";
 import { BUS_PASSENGER_TITLE_ITEMS } from "./constant";
 import * as S from "./style";
