+++ src/components/Member/MemberItem/index.tsx
@@ -1,15 +1,11 @@
-import React from "react";
 import { TBody } from "@b1nd/b1nd-dodamdodam-ui";
-import { useGetAllMemberListQuery } from "../../../queries/Member/Member.query";
+import { useGetAllMemberListQuery } from "queries/Member/Member.query";
 import * as S from "./style";
 import Student from "./Student";
 import Teacher from "./Teacher";
 import { useRecoilValue } from "recoil";
-import {
-  MemberSearch,
-  MemberSelectGrade,
-} from "../../../stores/Member/member.store";
-import { changeGrade } from "../../../utils/Member/changeGrade";
+import { MemberSearch, MemberSelectGrade } from "stores/Member/member.store";
+import { changeGrade } from "utils/Member/changeGrade";
 
 function MemberItem() {
   const { data: studentsInfo } = useGetAllMemberListQuery({
