+++ src/repositories/Member/MemberRepository.ts
@@ -1,4 +1,4 @@
-import { MemberListType } from "../../types/Member/Member.type";
+import { MemberListType } from "types/Member/Member.type";
 
 export interface MemberRepository {
   getAllMemberList(): Promise<MemberListType>;
