+++ src/utils/Member/sortStudentGrade.ts
@@ -1,4 +1,4 @@
-import { StudentType } from "../../types/Member/Member.type";
+import { StudentType } from "types/Member/Member.type";
 
 export const sortStudentGrade = (
   student1: StudentType,
